remove exclude: 'files' from default options
Re-importing a backup using the option exclude: 'files' in a Strapi instance where media is stored on a cloud provider removed all media from the cloud when on the import the files were not excluded as well. This can happen per accident why i removed it from the defaults. Better to export everything as default.